+++ packages/driver-typescript/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 - Initial release!
 - Added automatic project references support via the `--reference-workspaces` option.
-- Added a `typescript.create-project-config-file` event.
+- Added a `TypeScriptDriver#onCreateProjectConfigFile` event.