+++ packages/driver-typescript/CHANGELOG.md
@@ -1,7 +1,10 @@
-# 1.0.0
+# 1.0.0 - 2019-05-18
 
 #### 🎉 Release
 
 - Initial release!
+
+#### 🚀 Updates
+
 - Added automatic project references support via the `--reference-workspaces` option.
 - Added a `TypeScriptDriver#onCreateProjectConfigFile` event.